--- Proyecto_16_UpTask/classes/Emali.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace Classes;
-
-use PHPMailer\PHPMailer\PHPMailer;
-
-class Email
-{
-    protected $email;
-    protected $nombre;
-    protected $token;
-
-    public function __construct($email, $nombre, $token)
-    {
-        $this->email = $email;
-        $this->nombre = $nombre;
-        $this->token = $token;
-    }
-
-    public function enviarConfirmacion()
-    {
-
-        $mail = new PHPMailer();
-        $mail->isSMTP();
-        $mail->Host = 'sandbox.smtp.mailtrap.io';
-        $mail->SMTPAuth = true;
-        $mail->Port = 2525;
-        $mail->Username = 'b1ce5be9e2e4d6';
-        $mail->Password = '21a170ea5a025c';
-
-        $mail->setFrom('cuentas@uptask.com');
-        $mail->addAddress('cuentas@uptask.com', 'uptask.com');
-        $mail->Subject = 'Confirma tu Cuenta';
-
-        $mail->isHTML(TRUE);
-        $mail->CharSet = 'UTF-8';
-
-        $contenido = '<html>';
-        $contenido .= "<p><strong>Hola " . $this->nombre . "</strong> Has Creado tu cuenta en UpTask, solo debes confirmarla en el siguiente enlace</p>";
-        $contenido .= "<p>Presiona aquí: <a href='http://" . $_ENV['VHOST_NAME'] . " /confirmar?token=" . $this->token . "'>Confirmar Cuenta</a></p>";
-        $contenido .= "<p>Si tu no creaste esta cuenta, puedes ignorar este mensaje</p>";
-        $contenido .= '</html>';
-
-        $mail->Body = $contenido;
-
-        // Enviar el email
-        $mail->send();
-    }
-}
