+++ Proyecto_16_UpTask/includes/.env
@@ -4,3 +4,6 @@ DB_PASSWORD = ''
 DB_NAME = uptask
 DB_PORT = 3306
 VHOST_NAME = uptask.localhost
+MAILTRAP_HOST = sandbox.smtp.mailtrap.io
+MAILTRAP_USERNAME = b1ce5be9e2e4d6
+MAILTRAP_PASSWORD = 21a170ea5a025c
