+++ Proyecto_16_UpTask/public/index.php
@@ -16,6 +16,7 @@ $router->get("/crearCuenta", [LoginController::class, "crearCuenta"]);
 $router->post("/crearCuenta", [LoginController::class, "crearCuenta"]);
 
 $router->get("/mensajeConfirmarCuenta", [LoginController::class, "mensajeConfirmarCuenta"]);
+
 $router->get("/confirmarCuenta", [LoginController::class, "confirmarCuenta"]);
 
 $router->get("/olvidaPassword", [LoginController::class, "olvidaPassword"]);
